requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
httpx>=0.24.0
Pillow==10.2.0
pydantic==2.5.3
google-genai==0.2.2
python-dotenv==1.0.0
# Blockchain (Polygon) - Digital Certificate
web3==6.15.1
eth-account==0.11.0
# Supabase
supabase>=2.10.0
# Security
slowapi==0.1.9
firebase-admin==6.4.0
sqlalchemy>=2.0.0
